Another big commit.

+++ src/main/java/space/nixus/phoneduck/controller/ChannelController.java
@@ -1,10 +1,12 @@
 package space.nixus.phoneduck.controller;
 
-import space.nixus.phoneduck.error.AuthenticationException;
+import space.nixus.phoneduck.components.RadioHandler;
+import space.nixus.phoneduck.config.ApplicationConfig;
+import space.nixus.phoneduck.error.UnauthorizedException;
 import space.nixus.phoneduck.error.ChannelNotFoundException;
-import space.nixus.phoneduck.handler.RadioHandler;
+import space.nixus.phoneduck.error.PrivilegeException;
 import space.nixus.phoneduck.model.Channel;
-import space.nixus.phoneduck.model.Message;
+import space.nixus.phoneduck.model.ChannelParams;
 import space.nixus.phoneduck.service.RadioService;
 import space.nixus.phoneduck.service.UserService;
 import java.io.IOException;
@@ -12,12 +14,12 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.fasterxml.jackson.core.JsonProcessingException;
-import jakarta.websocket.server.PathParam;
 import space.nixus.phoneduck.utils.AuthHelper;
 
 @RestController
@@ -34,28 +36,31 @@ public class ChannelController {
 
     @GetMapping("/channels/")
     List<Channel> getChannels(
-            @RequestHeader("Authorization") String auth) throws AuthenticationException {
+            @RequestHeader(ApplicationConfig.AUTH_HEADER) String auth) throws UnauthorizedException {
         AuthHelper.checkAuth(userService, auth);
         return radioService.getChannels();
     }
 
     @PostMapping("/channels/")
     Channel addChannel(
-            @RequestHeader("Authorization") String auth,
-            @RequestParam("title") String title) throws AuthenticationException, JsonProcessingException, IOException {
+            @RequestHeader(ApplicationConfig.AUTH_HEADER) String auth,
+            @RequestBody ChannelParams params) throws UnauthorizedException, JsonProcessingException, IOException {
         var user = AuthHelper.checkAuth(userService, auth);
-        var channel = radioService.createChannel(user.getId(), title);
+        var channel = radioService.createChannel(user.getId(), params.getTitle());
         radioHandler.announce(channel);
         return channel;
     }
 
     @DeleteMapping("/channels/{id}")
     void removeChannel(
-            @RequestHeader("Authorization") String auth,
-            @PathParam("id") long id) throws AuthenticationException, ChannelNotFoundException {
-        AuthHelper.checkAuth(userService, auth);
-        radioService.removeChannel(id);
+            @RequestHeader(ApplicationConfig.AUTH_HEADER) String auth,
+            @PathVariable("id") Long id) throws UnauthorizedException, ChannelNotFoundException, PrivilegeException {
+        var user = AuthHelper.checkAuth(userService, auth);
+        var channel = radioService.getChannel(id);
+        if(user.getSuperuser() || user.getId().equals(channel.getOwnerId())) {
+            radioService.removeChannel(id);
+            return;
+        }
+        throw new PrivilegeException("Channel removal not allowed by this user.");
     }
-
-
 }
